requirements.txt
appdirs==1.4.4
asgiref==3.2.7
distlib==0.3.0
Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
filelock==3.0.12
importlib-metadata==1.6.0
importlib-resources==1.5.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.21
zipp==3.1.0
